requirements.txt
numpy
notebook
pandas
matplotlib
opencv-python
albumentations
ipykernel
tqdm
pycocotools
cython
from-root
boto3
mypy-boto3
Pillow
torch-summary
fastapi
uvicorn
Jinja2
python-multipart
PyYAML
dill==0.3.5.1
utils
torch>=1.7.0
torchvision>=0.8.1
python-box==6.0.2
ensure==1.0.2
-e .